Allows for the sorting of columns (#2)
* Allows for the sorting of columns

Previously, the columns would remain unsorted. This meant that the order
that they were created/specified in mysql was the order they appeared
in the file. This creates churn for developers that have loaded their
columns in different orders.

This is an optional setting because column order is not actually
portable if you do not specify the columns you are setting during an
update query.