Crude sanitizer for index values
Sorts index declarations and strips "USING BTREE", as it is the default
for innodb and myisam tables. Aborts when using any other engine, as
the defaults for in-memory tables and NDB are "USING HASH" instead.
Consider lookahead that checks the last line of the table definition as
a future improvement.